Begin work on unicode parsing script
Write a simple script that parses unicode decomposition data into tries
represented as C arrays. A big portion of the code was taken from a
version of the mkutf8data script by Olaf Weber [1], but adapted to work
with NFD, and to produce data structures more in line with those used
by hfsplus.

[1] https://marc.info/?l=linux-fsdevel&m=152584896931191&w=2