Run tests during build
Move the head of unicode.c and unicode.h to separate files, and write
alternative versions for running tests in user space. Since there are
no official unicode tests for case folding, the mktrie script must also
produce another version of the case folding trie, with no values.

Setup the Makefile so that the tests are run with every recompilation,
printing the results to build/test.out.